Fixing PrivateLinkItem.spec.js, TrashBin.spec.js
